moved middlewares and routers to individual files
