+++ api/db.mjs
@@ -21,11 +21,4 @@ const find = async () => {
   return await cursor.toArray();
 };
 
-const insertSignUp = async (insertData) => {
-  const db = await getDB();
-  const testCollection = db.collection("login");
-  const response = await testCollection.insertOne(insertData);
-  return response.insertedId;
-};
-
-export { insertOne, find, insertReview, insertSignUp };
+export { insertOne, find, insertReview };
