+++ routes/auth.mjs
@@ -1,0 +1,82 @@
+import { Router } from "express";
+import { requireAuthentication } from "../middleware.mjs";
+import {
+  register,
+  login,
+  logout,
+  sendPasswordResetLink,
+  resetAccountPassword,
+  refreshToken,
+  verifyAccountEmail,
+} from "../api/auth.mjs";
+
+const authRouter = Router();
+
+authRouter.post("/refreshToken", async (req, res) => {
+  try {
+    const rToken = req.body.refreshToken;
+    if (!rToken) {
+      return res.sendStatus(401);
+    }
+    res.json(await refreshToken(rToken));
+  } catch (e) {
+    console.error(e);
+    res.sendStatus(403);
+  }
+});
+
+authRouter.post("/register", async (req, res) => {
+  try {
+    await register(req.body);
+    res.json({ message: "account created" });
+  } catch (e) {
+    res.status(400);
+    res.json({ error: e });
+  }
+});
+
+authRouter.post("/login", async (req, res) => {
+  try {
+    const tokens = await login(req.body);
+    res.json(tokens);
+  } catch (e) {
+    res.status(400);
+    res.json({ error: e });
+  }
+});
+
+authRouter.post("/logout", requireAuthentication, async (req, res) => {
+  res.json(await logout(req.username));
+});
+
+authRouter.post("/requestPasswordReset", async (req, res) => {
+  try {
+    await sendPasswordResetLink(req.body);
+    res.sendStatus(200);
+  } catch (e) {
+    res.status(400);
+    res.json({ error: e });
+  }
+});
+
+authRouter.post("/resetAccountPassword", async (req, res) => {
+  try {
+    await resetAccountPassword(req.body);
+    res.sendStatus(200);
+  } catch (e) {
+    res.status(400);
+    res.json({ error: e });
+  }
+});
+
+authRouter.post("/verifyAccountEmail", async (req, res) => {
+  try {
+    await verifyAccountEmail(req.body);
+    res.sendStatus(200);
+  } catch (e) {
+    res.status(400);
+    res.json({ error: e });
+  }
+});
+
+export default authRouter;
