+++ routes/detail.mjs
@@ -1,0 +1,16 @@
+import { Router } from "express";
+import { detailLocation, detailStore } from "../api/detail.mjs";
+
+const detailRouter = Router();
+
+detailRouter.post("/location", async (req, res) => {
+  res.json(await detailLocation(req.body));
+});
+
+detailRouter.post("/store", async (req, res) => {
+  res.json(
+    await detailStore({ service: req.body.service, storeId: req.body.storeId })
+  );
+});
+
+export default detailRouter;
